requirements.txt
fastapi
uvicorn
httpx
beautifulsoup4
sentence-transformers
faiss-cpu
numpy
groq
python-dotenv
camel-tools
